# bpo-37358: Use vectorcall for functools.partial