--- a/Misc/NEWS.d/next/Library/2019-06-21-14-54-02.bpo-37358.RsASpn.rst
+++ b/Misc/NEWS.d/next/Library/2019-06-21-14-54-02.bpo-37358.RsASpn.rst
@@ -1,0 +1,1 @@
+Optimized ``functools.partial`` by using vectorcall.