@encukou 2019-07-05
Is `partial_vectorcall_no_args` useful? It kicks in only if no arguments (positional nor keyword) are added at all. Are such no-op partial objects worth optimizing?