@jdemeyer 2019-07-05
> Are such no-op partial objects worth optimizing?

To be honest, I have no idea. I just notices that this could be optimized, so I did it. The behaviour of `partial(func)` would be just like `staticmethod(func)` but callable itself.

Do you want me to remove that special case?