@encukou 2019-07-05
> To be honest, I have no idea. I just notices that this could be optimized, so I did it.

That sounds like premature optimization.
I wonder how much this PR itself helps; did you run any benchmarks?

> The behaviour of `partial(func)` would be just like `staticmethod(func)` but callable itself.

If that is an actual use case for that, just make `staticmethod` callable. AFAIK, in the discussions around that [⁽⁰⁾](https://mail.python.org/pipermail/python-ideas/2018-June/051585.html) [⁽¹⁾](https://bugs.python.org/issue20309), the only arguments against were not enough actual use cases.
We shouldn't optimize for people repurposing `partial` for this.

> Do you want me to remove that special case?

Yes, unless there is a need for the complication.