@jdemeyer 2019-07-08
I tried 3 simple micro-benchmarks (without PGO):

1. `f = partial(len)` and then benchmarking `f(())`
**before**: Mean +- std dev: 70.3 ns +- 0.2 ns
**after**: Mean +- std dev: 47.3 ns +- 0.1 ns
2. `f = partial(len,  [])` and then benchmarking `f()`
**before**: Mean +- std dev: 50.2 ns +- 0.2 ns
**after**: Mean +- std dev: 36.1 ns +- 0.3 ns
3. `f = partial(list.count, [])` and then benchmarking `f(0)`
**before**: Mean +- std dev: 84.4 ns +- 1.3 ns
**after**: Mean +- std dev: 47.9 ns +- 2.2 ns

So there is a clear performance improvement in all cases.