@jdemeyer 2019-07-08
Failed Travis CI build is due to `OSError: [Errno 113] No route to host` clearly not caused by this PR.